Fix out of range errors for malformed EIR data.
We've sometimes encountered in the wild devices that have malformed
EIR data structures.

For example, we've found the device 1c:52:16:b8:6a:f2 that
provided the `eir` buffer with data:
`[30,3,18,162,77,46,254,20,72,142,147,210,23,60,253,2,48,128,74,92,243,232,188,150,82,60,64,0,0,0,0]`.
This caused an out of bounds error on line 146, when
calling function `bytes.readUInt16LE(j)`
with `j = 28` which fails, since we're trying to read two bytes
starting from position 28 when `bytes` has only 29 elements.

The proposed solution respects the bounds of the `bytes` buffer when
reading from its positions directly.